pyproject.toml
[project]
name = "optimizador"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.119.1",
    "pulp>=3.3.0",
    "pydantic>=2.12.3",
    "python-jose[cryptography]>=3.5.0",
]
